+++ kernel/src/inc/usr/include/dev/storage/ata.h
@@ -26,7 +26,7 @@ extern u8 ata_read_buffer[512];
 void ata_init();
 
 void ata_read_one(u32 lba, u8* buffer);
-void ata_read_multiple(u32 lba, u8 sec_count, u8* buffer);
+void ata_read_multiple(u32 lba, u32 sec_count, u8* buffer);
 
 void ata_write_one(u32 lba, u8* buffer);
-void ata_write_multiple(u32 lba, u8 sec_count, u8* buffer);
+void ata_write_multiple(u32 lba, u32 sec_count, u8* buffer);
